refactor: PopulationTrendChartコンポーネント内でデータフォーマットさせていた部分を別のファイルに分離した。
