--- src/features/components/mockup.ts
@@ -1,294 +1,0 @@
-export const testPopulationApiData = {
-  message: null,
-  result: {
-    boundaryYear: 2020,
-    data: [
-      {
-        label: "総人口",
-        data: [
-          {
-            year: 1980,
-            value: 12817,
-          },
-          {
-            year: 1985,
-            value: 12707,
-          },
-          {
-            year: 1990,
-            value: 12571,
-          },
-          {
-            year: 1995,
-            value: 12602,
-          },
-          {
-            year: 2000,
-            value: 12199,
-          },
-          {
-            year: 2005,
-            value: 11518,
-          },
-          {
-            year: 2010,
-            value: 10888,
-          },
-          {
-            year: 2015,
-            value: 10133,
-          },
-          {
-            year: 2020,
-            value: 9302,
-          },
-          {
-            year: 2025,
-            value: 8431,
-          },
-          {
-            year: 2030,
-            value: 7610,
-          },
-          {
-            year: 2035,
-            value: 6816,
-          },
-          {
-            year: 2040,
-            value: 6048,
-          },
-          {
-            year: 2045,
-            value: 5324,
-          },
-        ],
-      },
-      {
-        label: "年少人口",
-        data: [
-          {
-            year: 1980,
-            value: 2906,
-            rate: 22.67,
-          },
-          {
-            year: 1985,
-            value: 2769,
-            rate: 21.79,
-          },
-          {
-            year: 1990,
-            value: 2346,
-            rate: 18.66,
-          },
-          {
-            year: 1995,
-            value: 2019,
-            rate: 16.02,
-          },
-          {
-            year: 2000,
-            value: 1728,
-            rate: 14.17,
-          },
-          {
-            year: 2005,
-            value: 1442,
-            rate: 12.52,
-          },
-          {
-            year: 2010,
-            value: 1321,
-            rate: 12.13,
-          },
-          {
-            year: 2015,
-            value: 1144,
-            rate: 11.29,
-          },
-          {
-            year: 2020,
-            value: 936,
-            rate: 10.06,
-          },
-          {
-            year: 2025,
-            value: 822,
-            rate: 9.75,
-          },
-          {
-            year: 2030,
-            value: 705,
-            rate: 9.26,
-          },
-          {
-            year: 2035,
-            value: 593,
-            rate: 8.7,
-          },
-          {
-            year: 2040,
-            value: 513,
-            rate: 8.48,
-          },
-          {
-            year: 2045,
-            value: 443,
-            rate: 8.32,
-          },
-        ],
-      },
-      {
-        label: "生産年齢人口",
-        data: [
-          {
-            year: 1980,
-            value: 8360,
-            rate: 65.23,
-          },
-          {
-            year: 1985,
-            value: 8236,
-            rate: 64.81,
-          },
-          {
-            year: 1990,
-            value: 8144,
-            rate: 64.78,
-          },
-          {
-            year: 1995,
-            value: 8048,
-            rate: 63.86,
-          },
-          {
-            year: 2000,
-            value: 7595,
-            rate: 62.26,
-          },
-          {
-            year: 2005,
-            value: 7032,
-            rate: 61.05,
-          },
-          {
-            year: 2010,
-            value: 6387,
-            rate: 58.66,
-          },
-          {
-            year: 2015,
-            value: 5538,
-            rate: 54.65,
-          },
-          {
-            year: 2020,
-            value: 4756,
-            rate: 51.13,
-          },
-          {
-            year: 2025,
-            value: 4187,
-            rate: 49.66,
-          },
-          {
-            year: 2030,
-            value: 3693,
-            rate: 48.53,
-          },
-          {
-            year: 2035,
-            value: 3251,
-            rate: 47.7,
-          },
-          {
-            year: 2040,
-            value: 2681,
-            rate: 44.33,
-          },
-          {
-            year: 2045,
-            value: 2261,
-            rate: 42.47,
-          },
-        ],
-      },
-      {
-        label: "老年人口",
-        data: [
-          {
-            year: 1980,
-            value: 1550,
-            rate: 12.09,
-          },
-          {
-            year: 1985,
-            value: 1702,
-            rate: 13.39,
-          },
-          {
-            year: 1990,
-            value: 2081,
-            rate: 16.55,
-          },
-          {
-            year: 1995,
-            value: 2535,
-            rate: 20.12,
-          },
-          {
-            year: 2000,
-            value: 2876,
-            rate: 23.58,
-          },
-          {
-            year: 2005,
-            value: 3044,
-            rate: 26.43,
-          },
-          {
-            year: 2010,
-            value: 3179,
-            rate: 29.2,
-          },
-          {
-            year: 2015,
-            value: 3442,
-            rate: 33.97,
-          },
-          {
-            year: 2020,
-            value: 3578,
-            rate: 38.46,
-          },
-          {
-            year: 2025,
-            value: 3422,
-            rate: 40.59,
-          },
-          {
-            year: 2030,
-            value: 3212,
-            rate: 42.21,
-          },
-          {
-            year: 2035,
-            value: 2972,
-            rate: 43.6,
-          },
-          {
-            year: 2040,
-            value: 2854,
-            rate: 47.19,
-          },
-          {
-            year: 2045,
-            value: 2620,
-            rate: 49.21,
-          },
-        ],
-      },
-    ],
-  },
-};
